refactor(scanner): extract secret patterns to satisfy quality ratchet
